feat: remove inner class
