+++ src/test/kotlin/smalltalk/support/EnableTestContainers.kt
@@ -3,12 +3,12 @@ package smalltalk.support
 import org.springframework.test.annotation.DirtiesContext
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.ContextConfiguration
-import smalltalk.support.testcontainers.TestContainersExtension
+import smalltalk.support.testcontainers.TestContainersInitializer
 
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
 @MustBeDocumented
-@ContextConfiguration(initializers = [TestContainersExtension.Initializer::class])
+@ContextConfiguration(initializers = [TestContainersInitializer::class])
 @ActiveProfiles("test")
 @DirtiesContext
-annotation class EnableTestContainers()
+annotation class EnableTestContainers
